fix: only attempt removal if exists

+++ packages/nx-plugin-openapi/src/utils/delete-output-dir.ts
@@ -1,6 +1,7 @@
 // From https://github.com/nrwl/nx/commit/7fcf136cad16a3111ffdc2965551f442e2618409#diff-0e0fabf373c42117939de9dac468f38475b2cb149476275279fecf66383cd2dc
 
 // Third Parties
+import { existsSync } from 'fs';
 import { resolve } from 'path';
 import * as rimraf from 'rimraf';
 
@@ -13,5 +14,7 @@ export function deleteOutputDir(root: string, outputPath: string) {
     throw new Error('Output path MUST not be project root directory!');
   }
 
-  rimraf.sync(resolvedOutputPath);
+  if (existsSync(resolvedOutputPath)) {
+    rimraf.sync(resolvedOutputPath);
+  }
 }
